requirements.txt
pygame-ce
https://github.com/umutcanekinci/pygame-core/zipball/master
pyyaml